requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
PyMuPDF
python-docx
sentence-transformers
faiss-cpu
numpy
openai
requests
transformers
accelerate
